\name{chr1ASW1000G}
\docType{data}
\alias{chr1ASW1000G}
\title{hapFabia Data: chr1ASW1000G genotype data in  \code{vcf} format}
\description{

  Genotype data in \code{vcf} format.
  Contains \code{chr1ASW1000G} of class character.

  A \code{vcf} file is produced by
  \code{write(chr1ASW1000G,file="chr1ASW1000G.vcf")}. 
  
}
  

\usage{chr1ASW1000G}
\format{  String \code{chr1ASW1000G} of class character.

}
\source{ 1000Genomes phase 1 release v3; chromosome 1, only ASW and only
few SNVs.}
\seealso{
\code{\link{HaploCluster-class}},
\code{\link{HaploClusterList-class}},
\code{\link{analyzeHaploClusters}},
\code{\link{compareHaploClusterLists}},
\code{\link{extractHaploClusters}},
\code{\link{findDenseRegions}},
\code{\link{hapFabia}},
\code{\link{hapFabiaVersion}},
\code{\link{hapRes}},
\code{\link{chr1ASW1000G}},
\code{\link{haploClusterList2excel}},
\code{\link{identifyDuplicates}},
\code{\link{iterateSegments}},
\code{\link{makePipelineFile}},
\code{\link{matrixPlot}},
\code{\link{mergeHaploClusterLists}},
\code{\link{mergedHaploClusterList}},
\code{\link{plotHaplotypeCluster}},
\code{\link{res}},
\code{\link{setAnnotation}},
\code{\link{setStatistics}},
\code{\link{sim}},
\code{\link{simu}},
\code{\link{simulateHaploClustersFabia}},
\code{\link{simulateHaplotypeClusters}},
\code{\link{split_sparse_matrix}},
\code{\link{toolsFactorizationClass}},
\code{\link{vcftoFABIA}}
}
\references{

S. Hochreiter et al.,
\sQuote{FABIA: Factor Analysis for Bicluster Acquisition},
Bioinformatics 26(12):1520-1527, 2010.

}
\keyword{datasets}


