\name{hapFabiaVersion}
\alias{hapFabiaVersion}
\title{Display version info for package \pkg{hapFabia}}
\description{
  \code{hapFabiaVersion} displays version information about the
  package.
}
\usage{
hapFabiaVersion()
}
\seealso{
\code{\link{HaploCluster-class}},
\code{\link{HaploClusterList-class}},
\code{\link{analyzeHaploClusters}},
\code{\link{compareHaploClusterLists}},
\code{\link{extractHaploClusters}},
\code{\link{findDenseRegions}},
\code{\link{hapFabia}},
\code{\link{hapFabiaVersion}},
\code{\link{hapRes}},
\code{\link{chr1ASW1000G}},
\code{\link{haploClusterList2excel}},
\code{\link{identifyDuplicates}},
\code{\link{iterateSegments}},
\code{\link{makePipelineFile}},
\code{\link{matrixPlot}},
\code{\link{mergeHaploClusterLists}},
\code{\link{mergedHaploClusterList}},
\code{\link{plotHaplotypeCluster}},
\code{\link{res}},
\code{\link{setAnnotation}},
\code{\link{setStatistics}},
\code{\link{sim}},
\code{\link{simu}},
\code{\link{simulateHaploClustersFabia}},
\code{\link{simulateHaplotypeClusters}},
\code{\link{split_sparse_matrix}},
\code{\link{toolsFactorizationClass}},
\code{\link{vcftoFABIA}}
}
\author{Sepp Hochreiter}
\examples{
hapFabiaVersion()
}
\keyword{models,multivariate,cluster}
\concept{genetics,haplotype,identity by descent,bicluster,next
  generation sequencing,genotype,single nucleotide polymorphism,single
  nucleotide variation,rare variants,rare SNPs, rare SNVs,rare
  haplotype clusters,short haplotype clusters, IBD}

