\name{mergedHaploClusterList}
\docType{data}
\alias{mergedHaploClusterList}

\title{hapFabia Data: \code{mergedHaploClusterList} is a result of
  \code{hapFabia} given as instance of the class \code{HaploClusterList}}

\description{

  Result of a \code{hapFabia} call
  given as instance of the class \code{HaploClusterList}.
  The object \code{mergedHaploClusterList}
  contains the extracted haplotype clusters.
  
}

\usage{mergedHaploClusterList}

\format{
  object \code{mergedHaploClusterList} of the class \code{HaploClusterList}
  
}
\source{result of a call of \code{hapFabia} }
\seealso{
\code{\link{HaploCluster-class}},
\code{\link{HaploClusterList-class}},
\code{\link{analyzeHaploClusters}},
\code{\link{compareHaploClusterLists}},
\code{\link{extractHaploClusters}},
\code{\link{findDenseRegions}},
\code{\link{hapFabia}},
\code{\link{hapFabiaVersion}},
\code{\link{hapRes}},
\code{\link{chr1ASW1000G}},
\code{\link{haploClusterList2excel}},
\code{\link{identifyDuplicates}},
\code{\link{iterateSegments}},
\code{\link{makePipelineFile}},
\code{\link{matrixPlot}},
\code{\link{mergeHaploClusterLists}},
\code{\link{mergedHaploClusterList}},
\code{\link{plotHaplotypeCluster}},
\code{\link{res}},
\code{\link{setAnnotation}},
\code{\link{setStatistics}},
\code{\link{sim}},
\code{\link{simu}},
\code{\link{simulateHaploClustersFabia}},
\code{\link{simulateHaplotypeClusters}},
\code{\link{split_sparse_matrix}},
\code{\link{toolsFactorizationClass}},
\code{\link{vcftoFABIA}}
}
\references{

S. Hochreiter et al.,
\sQuote{FABIA: Factor Analysis for Bicluster Acquisition},
Bioinformatics 26(12):1520-1527, 2010.

}
\keyword{datasets}


