\name{hapFabiaVersion}
\alias{hapFabiaVersion}
\title{Display version info for package \pkg{hapFabia}}
\description{
  \code{hapFabiaVersion} displays version information about the
  package.
}
\usage{
hapFabiaVersion()
}
\seealso{
\code{\link{IBDsegment-class}},
\code{\link{IBDsegmentList-class}},
\code{\link{analyzeIBDsegments}},
\code{\link{compareIBDsegmentLists}},
\code{\link{extractIBDsegments}},
\code{\link{findDenseRegions}},
\code{\link{hapFabia}},
\code{\link{hapFabiaVersion}},
\code{\link{hapRes}},
\code{\link{chr1ASW1000G}},
\code{\link{IBDsegmentList2excel}},
\code{\link{identifyDuplicates}},
\code{\link{iterateIntervals}},
\code{\link{makePipelineFile}},
\code{\link{matrixPlot}},
\code{\link{mergeIBDsegmentLists}},
\code{\link{mergedIBDsegmentList}},
\code{\link{plotIBDsegment}},
\code{\link{res}},
\code{\link{setAnnotation}},
\code{\link{setStatistics}},
\code{\link{sim}},
\code{\link{simu}},
\code{\link{simulateIBDsegmentsFabia}},
\code{\link{simulateIBDsegments}},
\code{\link{split_sparse_matrix}},
\code{\link{toolsFactorizationClass}},
\code{\link{vcftoFABIA}}
}
\author{Sepp Hochreiter}
\examples{
hapFabiaVersion()
}
\keyword{models,multivariate,cluster}
\concept{genetics,haplotype,identity by descent,bicluster,next
  generation sequencing,genotype,single nucleotide polymorphism,single
  nucleotide variation,rare variants,rare SNPs, rare SNVs,rare
  IBD segments,short IBD segments, IBD}

